+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/) and [Keep A Changelog's Format](http://keepachangelog.com/).
 
-## [Unreleased]
+## [2.32.0] 2018-11-07
 
 ### Added
 - `scrollTableOverflow` option for `<Table />`, this is now not enabled by default